+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `./default.nix` and `./flake.nix`.
+
 ## 2.3.1 - 2026-03-16
 
 ### Changed